requirements.txt
sentence-transformers
faiss-cpu
pandas
numpy
networkx
streamlit
